Fixed comments in scene load and unload
The main reason I replaced that code with comments is that it programatically wasn't making much sense, and it (sorta) made things harder to manage. And that isn't really what I'm going for here. I might implement some code that achieves the load/unload thing into the engine later on, though.